Refactor UI components and improve calculation precision
This commit refactors several UI components to use a more idiomatic callback pattern for state management and improves the precision of calculations by using `Double` instead of `Int` for intermediate values.

- In `components.kt`:
    - Renamed `sliderWithText`, `customCard2`, and `heightSliderWithText` to PascalCase (`SliderWithText`, `CustomCard2`, `HeightSliderWithText`).
    - Modified these components to use an `onValueChange: (String) -> Unit` callback instead of returning a value.
    - Updated `HeightSliderWithText` to use a `LaunchedEffect` for handling unit conversions.
    - Fixed a typo: renamed `preffix` parameter to `prefix`.
- Updated `BmiScreen.kt`, `EmiScreen.kt`, and `SipScreen.kt` to adapt to the new component signatures and naming conventions.
- In `BmiViewModel.kt`, `EmiViewModel.kt`, and `SipViewModel.kt`:
    - Switched from `toInt()` to `toDouble()` for input parsing to prevent precision loss during calculations.
    - Explicitly cast final results back to `Int` or formatted strings where required.
    - Forced `java.util.Locale.ENGLISH` in `String.format` within `BmiViewModel` for consistent decimal formatting.
- Updated validation logic in `bmiValidation.kt`, `EmiValidation.kt`, and `sipValidation.kt` to use `toDouble()` for range checks.
- Fixed a logic error in `sipValidation.kt` where the minimum Lumsum amount check was incorrectly compared against 100 instead of 500 (though the error message still says 500).

+++ app/src/main/java/com/plus/calculatorplus/presentation/validation/EmiValidation.kt
@@ -59,7 +59,7 @@ fun loanInterestRateValidation(
         }
 
         else -> {
-            val interestRateInt = interestRate.toInt()
+            val interestRateInt = interestRate.toDouble()
                if (interestRateInt < 0)  {
                 Pair(false, "Enter valid Interest Rate")
                 }
@@ -84,7 +84,7 @@ fun loanAmountValidation(
         }
 
         else -> {
-            val monthlyAmountInt = loanAmount.toInt()
+            val monthlyAmountInt = loanAmount.toDouble()
             if (monthlyAmountInt >= 10000) {
                 Pair(true, "")
             } else {
